jobs: run each job in separate thread
Instead of handling multiple consecutive jobs as a single entity, run
each job in a separate thread. Make sure that each followup job is
properly executed, by implementing a CONT action type that is really
just a placeholder. Rename cmd to job, so the vocabulary is clearer.

Also make sure to only run regexes once on each job. Add an origin flag
to the cli to be able to better track where a command originated from.