silent: add retry_sequence specific to silent jobs
Originally, the retry sequence was also a control for the number of
notifications on a flaky internet connection. For silent jobs the only
purpose remaining is resource conservation. Since e.g. checking for
internet every minute isn't expensive and we wouldn't want syncing to
halt for maybe another 15 minutes after the connection returning, it
would make sense to run such a silent job every minute at least. Add
a new retry_sequence_silent setting, which is used for silent jobs.