cli: split server and client
Currently, it is very easy to mix up arguments that are only for the
client and ones that are only for the server. The subparser solution
would only work it both server and client would have a subparser, but
that would not allow for prepending "throttle" to a cli command to send
it correctly to the server. Separate the two interface into two scripts,
leave throttle for the client and create throttle-server for the server.